+++ src/embed/VbrickEmbedConfig.ts
@@ -1,9 +1,12 @@
 import { VbrickSDKConfig } from '../VbrickSDK';
+import { VideoPlaybackSidebarButton } from './IVbrickTypes';
 import {PlaylistLayout} from './PlaylistLayout';
 
 /**
  * Options when creating the iframe embed for a video/webcast
  * @public
+ * @group Configuration
+ * @category Base
  */
 export interface VbrickBaseEmbedConfig extends VbrickSDKConfig {
 	/**
@@ -22,6 +25,13 @@ export interface VbrickBaseEmbedConfig extends VbrickSDKConfig {
 	height?: string;
 
 	/**
+	 * do not add the 'local-network-access' Permissions-Policy iframe "allow" attribute.
+	 * This may block ECDN / User Location Services functionality from working properly.
+	 * @default false
+	 */
+	noLocalNetworkAccess?: boolean;
+
+	/**
 	 * For video embeds. If a user needs to log in, go through the login process in a popup window. This is the standard behavior for non-SDK Rev embeded videos
 	 */
 	popupAuth?: boolean;
@@ -32,7 +42,7 @@ export interface VbrickBaseEmbedConfig extends VbrickSDKConfig {
 	timeoutSeconds?: number;
 
 	autoplay?: boolean;
-	
+
 	/**
 	 * set the volume to upon initial load (for muting or otherwise)
 	 */
@@ -42,10 +52,15 @@ export interface VbrickBaseEmbedConfig extends VbrickSDKConfig {
 /**
  * Options available when embedding a VOD/video
  * @public
+ * @group Configuration
+ * @category VOD
  */
 export interface VbrickVideoEmbedConfig extends VbrickBaseEmbedConfig {
 	playInLoop?: boolean;
 	hideChapters?: boolean;
+	hideChapterNavigation?: boolean;
+	hideChapterMenu?: boolean;
+	hideChapterImages?: boolean;
 	hideOverlayControls?: boolean;
 	hidePlayControls?: boolean;
 	hideSubtitles?: boolean;
@@ -53,6 +68,40 @@ export interface VbrickVideoEmbedConfig extends VbrickBaseEmbedConfig {
 	forcedCaptions?: boolean;
 	hideSettings?: boolean;
 	hideFullscreen?: boolean;
+
+	/**
+	 * shows video player with sidebar.
+	 */
+	showFullPlayer?: boolean;
+
+	/**
+	 * provide the sidebar name to be opened by default when video loads.
+	 * showFullPlayer value needs to be true for this to work.
+	 */
+	defaultSidebar?: VideoPlaybackSidebarButton;
+
+	/**
+	 * query to filter sidebar items. esp. use in pulse sidebar.
+	 * showFullPlayer = true and defaultSidebar should have a value to make it work.
+	 */
+	sidebarFilterQuery?: string;
+	/**
+	 * flag to override account theme to default theme.
+	 */
+	applyDefaultTheme?: boolean;
+
+	/**
+	 * In case of showFullPlayer = true, user can provide a mobile breakpoint size
+	 * so that player and sidebar will render nicely together.
+	 * if not provided then 991 is the default value for mobile layout.
+	 */
+	mobileLayoutBreakPoint?: number;
+
+	/**
+	 * In case of showFullPlayer = true, user can customize the sidebar tabs to include
+	 * If not provided or undefined then the default value for each tab is true
+	 */
+	sidebarTabs?: Partial<Record<`${VideoPlaybackSidebarButton}`, boolean>>
 
 	/**
 	 * Starts the video at specified timestamp. must be in the format ##m##s. For example 00m30s.
@@ -87,15 +136,37 @@ export interface VbrickVideoEmbedConfig extends VbrickBaseEmbedConfig {
 	noPlayBar?: boolean;
 	/** @deprecated - embed parameter name - alias of hideSettings */
 	noSettings?: boolean;
+	/** @deprecated - embed parameter name - alias of hideChapterNavigation */
+	noChapterSeek?: boolean;
+	/** @deprecated - embed parameter name - alias of hideChapterImages */
+	noChapterDisplay?: boolean;
+	/** @deprecated - embed parameter name - alias of hideChapterMenu */
+	noChapterMenu?: boolean;
+
+	/** @deprecated - embed parameter name - alias for sidebar tab option */
+	hideInfo?: boolean;
+	/** @deprecated - embed parameter name - alias for sidebar tab option */
+	hideComments?: boolean;
+	/** @deprecated - embed parameter name - alias for sidebar tab option */
+	hidePulse?: boolean;
+	/** @deprecated - embed parameter name - alias for sidebar tab option */
+	hideReview?: boolean;
+	/** @deprecated - embed parameter name - alias for sidebar tab option */
+	hidePlaylist?: boolean;
+	// hideChapters?: boolean;
+	/** @deprecated - embed parameter name - alias for sidebar tab option */
+	hideAnalytics?: boolean;
 }
 
 /**
  * Options available when embedding a webcast
  * @public
+ * @group Configuration
+ * @category Webcast
  */
 export interface VbrickWebcastEmbedConfig extends VbrickBaseEmbedConfig {
 	/**
-	 * Include Chat, QA and Polls widgets in embed (if configured) 
+	 * Include Chat, QA and Polls widgets in embed (if configured)
 	 */
 	showFullWebcast?: boolean;
 
@@ -105,6 +176,8 @@ export interface VbrickWebcastEmbedConfig extends VbrickBaseEmbedConfig {
 
 /**
  * @public
+ * @group Configuration
+ * @category Playlist
  */
 export interface VbrickPlaylistEmbedConfig extends VbrickVideoEmbedConfig {
 	/**
@@ -135,5 +208,7 @@ export interface VbrickPlaylistEmbedConfig extends VbrickVideoEmbedConfig {
 /**
  * Options available when embedding a VOD/video or webcast
  * @public
+ * @group Configuration
+ * @category Base
  */
 export interface VbrickEmbedConfig extends VbrickVideoEmbedConfig, VbrickWebcastEmbedConfig, VbrickPlaylistEmbedConfig {}